Backend/app/core/config.py
from __future__ import annotations

import os
from pathlib import Path
from dataclasses import dataclass

from dotenv import load_dotenv


_backend_dir = Path(__file__).resolve().parents[2]
load_dotenv(_backend_dir / ".env")


@dataclass(frozen=True)
class Settings:
    database_url: str = os.getenv(
        "DATABASE_URL",
        "postgresql+psycopg://postgres:postgres@localhost:5432/dodgeai",
    )

    openai_api_key: str | None = os.getenv("OPENAI_API_KEY") or None
    openai_model: str = os.getenv("OPENAI_MODEL", "gpt-4o-mini")

    log_level: str = os.getenv("LOG_LEVEL", "INFO")


settings = Settings()

